Move ContactDetails and NamedEntity to facade.common
